Fix broken links to blog posts
Fix issue in which the blog post list on any page but the 1st would
include broken links to blog posts.

Links to blog posts are now relative to the site URL rather than the
current URL so that pagination elements in the URL (e.g., /pages/n)
are not prepended to the blog post links.